renamed data directory to pocket_friends to avoid conflict
