+++ pocket_friends.py
@@ -3,8 +3,8 @@ Launch script for Pocket Friends.
 """
 import pygame
 import sys
-from data.game import main as game_main
-from data.development.dev_menu import main as dev_menu_main
+from pocket_friends.game import main as game_main
+from pocket_friends.development.dev_menu import main as dev_menu_main
 
 enable_dev = False
