+++ pocket_friends/development/dev_menu.py
@@ -1,7 +1,7 @@
 """
 Development menu for the game on Raspberry Pi. NOTE: THIS DOES NOTHING ON A COMPUTER!
 """
-import data.game
+import pocket_friends.game
 import importlib.util
 import os
 import pygame
@@ -16,7 +16,7 @@ try:
     importlib.util.find_spec('RPi.GPIO')
     import RPi.GPIO as GPIO
 except ImportError:
-    import data.development.FakeGPIO as GPIO
+    import pocket_friends.development.FakeGPIO as GPIO
 
 # Global variable to keep track of the current menu.
 menu = 'main'
@@ -43,7 +43,7 @@ def start_game():
     Cleans the GPIO and starts the game.
     """
     GPIOHandler.teardown()
-    data.game.main()
+    pocket_friends.game.main()
     pygame.quit()
     GPIOHandler.setup()
 
@@ -95,7 +95,7 @@ def main():
 
     # The following defines all of the options in the various different menus.
 
-    main_menu = Menu('Pocket Friends Dev Menu {0}\nGame Version {1}'.format(dev_version, data.game.version))
+    main_menu = Menu('Pocket Friends Dev Menu {0}\nGame Version {1}'.format(dev_version, pocket_friends.game.version))
     main_menu.add_option(Menu.Option('Start Game', start_game))
     main_menu.add_option(Menu.Option('Button Test', run_button_test))
     main_menu.add_option(Menu.Option('Restart Dev Menu', quit_with_error))
